requirements.txt
feedparser==6.0.2
azure-devops==6.0.0b4
